requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.3
sentence-transformers==2.3.1
torch==2.1.2
scikit-learn==1.4.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
black==23.12.0
flake8==6.1.0
moviepy==1.0.3
nest-asyncio==1.6.0
tqdm==4.66.1
dashscope>=1.10.0  # 阿里云DashScope API
